Move Ember.typeOf and Ember.isArray back into ember-runtime.
These should have never been moved into ember-metal as they depend on ember-runtime.

The uses of them in ember-metal was few and dubious.
