+++ packages/ember-metal/lib/expand_properties.js
@@ -1,6 +1,5 @@
 import EmberError from 'ember-metal/error';
-import { forEach } from 'ember-metal/enumerable_utils';
-import { typeOf } from 'ember-metal/utils';
+import { forEach } from 'ember-metal/array';
 
 /**
   @module ember-metal
@@ -39,17 +38,17 @@ export default function expandProperties(pattern, callback) {
     throw new EmberError(`Brace expanded properties cannot contain spaces, e.g. 'user.{firstName, lastName}' should be 'user.{firstName,lastName}'`);
   }
 
-  if ('string' === typeOf(pattern)) {
+  if ('string' === typeof pattern) {
     var parts = pattern.split(SPLIT_REGEX);
     var properties = [parts];
 
-    forEach(parts, function(part, index) {
+    forEach.call(parts, function(part, index) {
       if (part.indexOf(',') >= 0) {
         properties = duplicateAndReplace(properties, part.split(','), index);
       }
     });
 
-    forEach(properties, function(property) {
+    forEach.call(properties, function(property) {
       callback(property.join(''));
     });
   } else {
@@ -60,8 +59,8 @@ export default function expandProperties(pattern, callback) {
 function duplicateAndReplace(properties, currentParts, index) {
   var all = [];
 
-  forEach(properties, (property) => {
-    forEach(currentParts, (part) => {
+  forEach.call(properties, (property) => {
+    forEach.call(currentParts, (part) => {
       var current = property.slice(0);
       current[index] = part;
       all.push(current);
