+++ packages/ember-metal/lib/watching.js
@@ -3,7 +3,7 @@
 */
 
 import {
-  typeOf
+  isArray
 } from "ember-metal/utils";
 import {
   removeChainWatcher,
@@ -37,7 +37,7 @@ import {
 */
 function watch(obj, _keyPath, m) {
   // can't watch length on Array - it is special...
-  if (_keyPath === 'length' && typeOf(obj) === 'array') { return; }
+  if (_keyPath === 'length' && isArray(obj)) { return; }
 
   if (!isPath(_keyPath)) {
     watchKey(obj, _keyPath, m);
@@ -57,7 +57,7 @@ watch.flushPending = flushPendingChains;
 
 export function unwatch(obj, _keyPath, m) {
   // can't watch length on Array - it is special...
-  if (_keyPath === 'length' && typeOf(obj) === 'array') { return; }
+  if (_keyPath === 'length' && isArray(obj)) { return; }
 
   if (!isPath(_keyPath)) {
     unwatchKey(obj, _keyPath, m);
