+++ packages/ember-routing/lib/system/generate_controller.js
@@ -1,6 +1,6 @@
 import Ember from "ember-metal/core"; // Logger
 import { get } from "ember-metal/property_get";
-import { isArray } from "ember-metal/utils";
+import { isArray } from "ember-runtime/utils";
 
 /**
 @module ember
