+++ packages/ember-runtime/lib/main.js
@@ -87,6 +87,8 @@ import Service from 'ember-runtime/system/service';
 import RSVP from 'ember-runtime/ext/rsvp';     // just for side effect of extending Ember.RSVP
 import 'ember-runtime/ext/string';   // just for side effect of extending String.prototype
 import 'ember-runtime/ext/function'; // just for side effect of extending Function.prototype
+
+import { typeOf, isArray } from 'ember-runtime/utils';
 // END IMPORTS
 
 // BEGIN EXPORTS
@@ -122,6 +124,9 @@ Ember.arrayComputed = arrayComputed;
 Ember.ArrayComputedProperty = ArrayComputedProperty;
 Ember.reduceComputed = reduceComputed;
 Ember.ReduceComputedProperty = ReduceComputedProperty;
+
+Ember.typeOf    = typeOf;
+Ember.isArray   = isArray;
 
 // ES6TODO: this seems a less than ideal way/place to add properties to Ember.computed
 var EmComputed = Ember.computed;
