+++ packages/ember-views/lib/mixins/class_names_support.js
@@ -16,7 +16,7 @@ import {
 } from "ember-metal/streams/utils";
 import { streamifyClassNameBinding } from "ember-views/streams/class_name_binding";
 import {
-  typeOf
+  isArray
 } from "ember-metal/utils";
 
 var EMPTY_ARRAY = [];
@@ -31,10 +31,10 @@ var ClassNamesSupport = Mixin.create({
   init() {
     this._super(...arguments);
 
-    Ember.assert("Only arrays are allowed for 'classNameBindings'", typeOf(this.classNameBindings) === 'array');
+    Ember.assert("Only arrays are allowed for 'classNameBindings'", isArray(this.classNameBindings));
     this.classNameBindings = emberA(this.classNameBindings.slice());
 
-    Ember.assert("Only arrays of static class strings are allowed for 'classNames'. For dynamic classes, use 'classNameBindings'.", typeOf(this.classNames) === 'array');
+    Ember.assert("Only arrays of static class strings are allowed for 'classNames'. For dynamic classes, use 'classNameBindings'.", isArray(this.classNames));
     this.classNames = emberA(this.classNames.slice());
   },
