+++ packages/ember-views/lib/views/select.js
@@ -15,7 +15,7 @@ import { get } from "ember-metal/property_get";
 import { set } from "ember-metal/property_set";
 import View from "ember-views/views/view";
 import CollectionView from "ember-views/views/collection_view";
-import { isArray } from "ember-metal/utils";
+import { isArray } from "ember-runtime/utils";
 import isNone from 'ember-metal/is_none';
 import { computed } from "ember-metal/computed";
 import { A as emberA } from "ember-runtime/system/native_array";
